requirements.txt
langchain
langchain-classic
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
langchain-chroma
langchain-ollama
chromadb
sentence-transformers
python-dotenv
streamlit
streamlit-chat
ragas
atlassian-python-api
pypdf
pandas
rank_bm25
